fix: fix asset links after refactor
