+++ apps/guardian-ui/src/components/setup/screens/verifyGuardians/VerifyGuardians.tsx
@@ -32,11 +32,11 @@ import { useTranslation } from '@fedimint/utils';
 import { CopyInput, Table } from '@fedimint/ui';
 import { useConsensusPolling, useSetupContext } from '../../../../hooks';
 import { GuardianRole } from '../../../../types';
-import { ReactComponent as ArrowRightIcon } from '../assets/svgs/arrow-right.svg';
-import { ReactComponent as CopyIcon } from '../assets/svgs/copy.svg';
+import { ReactComponent as ArrowRightIcon } from '../../../../assets/svgs/arrow-right.svg';
+import { ReactComponent as CopyIcon } from '../../../../assets/svgs/copy.svg';
 import { formatApiErrorMessage } from '../../../../utils/api';
-import { ReactComponent as CheckCircleIcon } from '../assets/svgs/check-circle.svg';
-import { ReactComponent as XCircleIcon } from '../assets/svgs/x-circle.svg';
+import { ReactComponent as CheckCircleIcon } from '../../../../assets/svgs/check-circle.svg';
+import { ReactComponent as XCircleIcon } from '../../../../assets/svgs/x-circle.svg';
 
 interface PeerWithHash {
   id: string;
